react-i18n: Add support for self-closing elements in formatHTML (#7034)
* Split, convert react-i18n package to TypeScript

* react-i18n: Add support for self-closing elements in formatHTML

changelog: Internal, Localization, Support self-closing HTML tags in localization helper

* Add documentation for react-i18n

* Run Prettier on README

* Add (failing) regression coverage for dropped text

See: https://github.com/18F/identity-idp/pull/7034#discussion_r981716369

* Fix dropped text following self-closing tag
